sigmaops/requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.25.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
Faker>=20.0.0
streamlit-option-menu>=0.3.12
